+++ src/models/Color.js
@@ -6,5 +6,9 @@ export default {
   loss: '#ca0813',
   draw: '#fed86f',
   darkThemeText: '#d1d5da',
-  lightThemeText: '#666'
+  lightThemeText: '#666',
+  ally: '#299FFD',
+  transparentAlly: 'rgba(41,159,253,0.7)',
+  enemy: '#FD9629',
+  transparentEnemy: 'rgba(253,150,41,0.7)'
 }