+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.1.2
 
+- Added a 'Trends' page with charts about your match results
+
 ## Version 0.1.1
 
 - Fixed 'delete match' not working [#11](https://github.com/cheshire137/competiwatch-desktop/issues/11)